pyproject.toml
[tool.poetry]
name = "AutodocRemote"
version = "0.1.0"
description = ""
authors = ["Tesla2000 <fratajczak124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.7.1"
dynamic-executor = "^1.2.6"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
libcst = "^1.4.0"
litellm = "^1.46.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autodocremote = "AutodocRemote:main"
